+++ src/pages/test-report/__tests__/test-report.analytics.effects.spec.ts
@@ -34,7 +34,7 @@ import { testReportPracticeModeSlot } from '../../../modules/tests/__mocks__/tes
 import {
   manoeuvreTypeLabels,
   manoeuvreCompetencyLabels,
-} from '../components/manoeuvre-competency/manoeuvre-competency.constants';
+} from '../cat-b/components/manoeuvre-competency/manoeuvre-competency.constants';
 import { AnalyticRecorded } from '../../../providers/analytics/analytics.actions';
 import { legalRequirementsLabels, legalRequirementToggleValues }
   from '../../../shared/constants/legal-requirements/catb-legal-requirements';
