+++ src/pages/test-report/cat-b/__tests__/test-report.cat-b.page.spec.ts
@@ -1,6 +1,6 @@
 
-import { ManoeuvresComponent } from '../../components/manoeuvres/manoeuvres';
-import { ManoeuvresPopoverComponent } from '../../components/manoeuvres-popover/manoeuvres-popover';
+import { ManoeuvresComponent } from '../components/manoeuvres/manoeuvres';
+import { ManoeuvresPopoverComponent } from '../components/manoeuvres-popover/manoeuvres-popover';
 import { ComponentFixture, async, TestBed } from '@angular/core/testing';
 import { IonicModule, NavController, NavParams, Config, Platform, ModalController } from 'ionic-angular';
 import {
@@ -30,9 +30,9 @@ import { testReportReducer } from '../../test-report.reducer';
 import { LegalRequirementComponent } from '../../components/legal-requirement/legal-requirement';
 import { EtaComponent } from '../../components/examiner-takes-action/eta';
 import { initialState } from '../../../../modules/tests/test-data/test-data.reducer';
-import { ControlledStopComponent } from '../../components/controlled-stop/controlled-stop';
-import { ManoeuvreCompetencyComponent } from '../../components/manoeuvre-competency/manoeuvre-competency';
-import { VehicleCheckComponent } from '../../components/vehicle-check/vehicle-check';
+import { ControlledStopComponent } from '../components/controlled-stop/controlled-stop';
+import { ManoeuvreCompetencyComponent } from '../components/manoeuvre-competency/manoeuvre-competency';
+import { VehicleCheckComponent } from '../components/vehicle-check/vehicle-check';
 import { EcoComponent } from '../../components/eco/eco';
 import { TestReportValidatorProvider } from '../../../../providers/test-report-validator/test-report-validator';
 import {
