+++ src/pages/test-report/cat-b/components/controlled-stop/__tests__/controlled-stop.spec.ts
@@ -3,25 +3,27 @@ import { By } from '@angular/platform-browser';
 import { ControlledStopComponent } from '../controlled-stop';
 import { IonicModule } from 'ionic-angular';
 import { StoreModule, Store } from '@ngrx/store';
-import { testsReducer } from '../../../../../modules/tests/tests.reducer';
-import { testReportReducer } from '../../../test-report.reducer';
-import { StoreModel } from '../../../../../shared/models/store.model';
+import { testsReducer } from '../../../../../../modules/tests/tests.reducer';
+import { testReportReducer } from '../../../../test-report.reducer';
+import { StoreModel } from '../../../../../../shared/models/store.model';
 import { MockComponent } from 'ng-mocks';
-import { CompetencyButtonComponent } from '../../../components/competency-button/competency-button';
-import { TickIndicatorComponent } from '../../../../../components/common/tick-indicator/tick-indicator';
+import { CompetencyButtonComponent } from '../../../../components/competency-button/competency-button';
+import { TickIndicatorComponent } from '../../../../../../components/common/tick-indicator/tick-indicator';
 import {
-  DrivingFaultsBadgeComponent } from '../../../../../components/common/driving-faults-badge/driving-faults-badge';
-import { SeriousFaultBadgeComponent } from '../../../../../components/common/serious-fault-badge/serious-fault-badge';
+  DrivingFaultsBadgeComponent } from '../../../../../../components/common/driving-faults-badge/driving-faults-badge';
+import {
+  SeriousFaultBadgeComponent,
+} from '../../../../../../components/common/serious-fault-badge/serious-fault-badge';
 import {
   DangerousFaultBadgeComponent,
-} from '../../../../../components/common/dangerous-fault-badge/dangerous-fault-badge';
-import { StartTest } from '../../../../../modules/tests/tests.actions';
+} from '../../../../../../components/common/dangerous-fault-badge/dangerous-fault-badge';
+import { StartTest } from '../../../../../../modules/tests/tests.actions';
 import {
   ControlledStopAddDrivingFault,
   ControlledStopRemoveFault,
-} from '../../../../../modules/tests/test-data/controlled-stop/controlled-stop.actions';
-import { CompetencyOutcome } from '../../../../../shared/models/competency-outcome';
-import { TestCategory } from '../../../../../shared/models/test-category';
+} from '../../../../../../modules/tests/test-data/controlled-stop/controlled-stop.actions';
+import { CompetencyOutcome } from '../../../../../../shared/models/competency-outcome';
+import { TestCategory } from '../../../../../../shared/models/test-category';
 
 describe('ControlledStopComponent', () => {
   let fixture: ComponentFixture<ControlledStopComponent>;
