+++ src/pages/test-report/cat-b/components/controlled-stop/controlled-stop.scss
