+++ src/pages/test-report/cat-b/components/controlled-stop/controlled-stop.ts
@@ -4,25 +4,25 @@ import { Subscription } from 'rxjs/Subscription';
 import { merge } from 'rxjs/observable/merge';
 import { map } from 'rxjs/operators';
 
-import { StoreModel } from '../../../../shared/models/store.model';
+import { StoreModel } from '../../../../../shared/models/store.model';
 import { Store, select } from '@ngrx/store';
-import { getTests } from '../../../../modules/tests/tests.reducer';
-import { getCurrentTest } from '../../../../modules/tests/tests.selector';
-import { getTestData } from '../../../../modules/tests/test-data/test-data.reducer';
+import { getTests } from '../../../../../modules/tests/tests.reducer';
+import { getCurrentTest } from '../../../../../modules/tests/tests.selector';
+import { getTestData } from '../../../../../modules/tests/test-data/test-data.reducer';
 import {
   hasControlledStopBeenCompleted,
-} from '../../../../modules/tests/test-data/test-data.selector';
+} from '../../../../../modules/tests/test-data/test-data.selector';
 import {
   ToggleControlledStop,
   ControlledStopAddDangerousFault,
   ControlledStopAddSeriousFault,
   ControlledStopAddDrivingFault,
   ControlledStopRemoveFault,
-} from '../../../../modules/tests/test-data/controlled-stop/controlled-stop.actions';
-import { getTestReportState } from '../../test-report.reducer';
-import { isRemoveFaultMode, isSeriousMode, isDangerousMode } from '../../test-report.selector';
-import { ToggleRemoveFaultMode, ToggleSeriousFaultMode, ToggleDangerousFaultMode } from '../../test-report.actions';
-import { CompetencyOutcome } from '../../../../shared/models/competency-outcome';
+} from '../../../../../modules/tests/test-data/controlled-stop/controlled-stop.actions';
+import { getTestReportState } from '../../../test-report.reducer';
+import { isRemoveFaultMode, isSeriousMode, isDangerousMode } from '../../../test-report.selector';
+import { ToggleRemoveFaultMode, ToggleSeriousFaultMode, ToggleDangerousFaultMode } from '../../../test-report.actions';
+import { CompetencyOutcome } from '../../../../../shared/models/competency-outcome';
 import { get } from 'lodash';
 
 interface ControlledStopComponentState {
