+++ src/pages/test-report/cat-b/components/manoeuvre-competency/manoeuvre-competency.html
