+++ src/pages/test-report/cat-b/components/manoeuvres-popover/manoeuvres-popover.scss
