+++ src/pages/test-report/cat-b/components/manoeuvres/manoeuvres.scss
