+++ src/pages/test-report/cat-b/components/vehicle-check/vehicle-check.scss
