+++ src/pages/test-report/components/test-report-components.module.ts
@@ -4,8 +4,6 @@ import { CommonModule } from '@angular/common';
 import { IonicModule } from 'ionic-angular';
 
 import { EtaComponent } from './examiner-takes-action/eta';
-import { ManoeuvresPopoverComponent } from './manoeuvres-popover/manoeuvres-popover';
-import { ManoeuvresComponent } from './manoeuvres/manoeuvres';
 import { CompetencyComponent } from './competency/competency';
 import { CompetencyButtonComponent } from './competency-button/competency-button';
 import { DrivingFaultSummaryComponent } from './driving-fault-summary/driving-fault-summary';
@@ -13,9 +11,6 @@ import { ToolbarComponent } from './toolbar/toolbar';
 import { SeriousTooltipComponent } from './serious-tooltip/serious-tooltip';
 import { DangerousTooltipComponent } from './dangerous-tooltip/dangerous-tooltip';
 import { LegalRequirementComponent } from './legal-requirement/legal-requirement';
-import { ControlledStopComponent } from './controlled-stop/controlled-stop';
-import { ManoeuvreCompetencyComponent } from './manoeuvre-competency/manoeuvre-competency';
-import { VehicleCheckComponent } from './vehicle-check/vehicle-check';
 import { EcoComponent } from './eco/eco';
 import { TimerComponent } from './timer/timer';
 import { EndTestModalModule } from './end-test-modal/end-test-modal.module';
@@ -25,8 +20,6 @@ import { EtaInvalidModalModule } from './eta-invalid-modal/eta-invalid-modal.mod
 @NgModule({
   declarations: [
     EtaComponent,
-    ManoeuvresPopoverComponent,
-    ManoeuvresComponent,
     CompetencyComponent,
     CompetencyButtonComponent,
     DrivingFaultSummaryComponent,
@@ -34,9 +27,6 @@ import { EtaInvalidModalModule } from './eta-invalid-modal/eta-invalid-modal.mod
     SeriousTooltipComponent,
     DangerousTooltipComponent,
     LegalRequirementComponent,
-    ControlledStopComponent,
-    ManoeuvreCompetencyComponent,
-    VehicleCheckComponent,
     EcoComponent,
     TimerComponent,
   ],
@@ -50,8 +40,6 @@ import { EtaInvalidModalModule } from './eta-invalid-modal/eta-invalid-modal.mod
   ],
   exports:[
     EtaComponent,
-    ManoeuvresPopoverComponent,
-    ManoeuvresComponent,
     CompetencyComponent,
     CompetencyButtonComponent,
     DrivingFaultSummaryComponent,
@@ -59,9 +47,6 @@ import { EtaInvalidModalModule } from './eta-invalid-modal/eta-invalid-modal.mod
     SeriousTooltipComponent,
     DangerousTooltipComponent,
     LegalRequirementComponent,
-    ControlledStopComponent,
-    ManoeuvreCompetencyComponent,
-    VehicleCheckComponent,
     EcoComponent,
     TimerComponent,
   ],
