+++ src/pages/view-test-result/__tests__/view-test-result.spec.ts
@@ -39,7 +39,9 @@ import { ViewTestHeaderComponent } from '../components/view-test-header/view-tes
 import { ViewTestHeaderModel } from '../components/view-test-header/view-test-header.model';
 import { TestOutcome } from '../../../modules/tests/tests.constants';
 import { DebriefCardComponent } from '../components/debrief-card/debrief-card';
-import { manoeuvreTypeLabels } from '../../test-report/components/manoeuvre-competency/manoeuvre-competency.constants';
+import {
+  manoeuvreTypeLabels,
+} from '../../test-report/cat-b/components/manoeuvre-competency/manoeuvre-competency.constants';
 import { DebriefCardModel } from '../components/debrief-card/debrief-card.model';
 import { CompetencyOutcome } from '../../../shared/models/competency-outcome';
 import { ErrorMessageComponent } from '../../../components/common/error-message/error-message';
